refactor(background): extract lifecycle modules (omnibox, port-messaging, commands-listener)
Move omnibox integration, quick-search port messaging, keyboard shortcuts, and keep-alive alarms into lifecycle/*.ts modules. Service-worker.ts shrinks from 473 to 192 lines — pure init orchestration and message dispatch.

Made-with: Cursor
